Improve pipeline stages ast errors.
They would throw 500 instead of 4xx before.
