+++ pkg/logql/evaluator.go
@@ -435,33 +435,6 @@ type rangeVectorEvaluator struct {
 	err error
 }
 
-type pipelineError struct {
-	metric    labels.Labels
-	errorType string
-}
-
-func newPipelineErr(metric labels.Labels) *pipelineError {
-	return &pipelineError{
-		metric:    metric,
-		errorType: metric.Get(log.ErrorLabel),
-	}
-}
-
-func (e pipelineError) Error() string {
-	return fmt.Sprintf(
-		"pipeline error: '%s' for series: '%s'.\n"+
-			"Use a label filter to intentionally skip this error. (e.g | __error__!=\"%s\").\n"+
-			"To skip all potential errors you can match empty errors.(e.g __error__=\"\")\n"+
-			"The label filter can also be specified after unwrap. (e.g | unwrap latency | __error__=\"\" )\n",
-		e.errorType, e.metric, e.errorType)
-}
-
-// IsPipelineError tells if the error is generated by a Pipeline.
-func IsPipelineError(err error) bool {
-	_, ok := err.(*pipelineError)
-	return ok
-}
-
 func (r *rangeVectorEvaluator) Next() (bool, int64, promql.Vector) {
 	next := r.iter.Next()
 	if !next {
